(TASKS-19) Create CLI accepting hosts, user, password and command
Adds a CLI using trollop. It accepts one or more space separated hosts. Each
host can be specified as a URI, e.g. ssh://host:port or winrm://host:port. The
port will default based on the scheme, and we default to ssh if no scheme is
given.

Also adds --user and --password options. They are required for winrm, but not
necessarily for ssh, depending on your ssh config.

Also adds the ability to specify task options, eg key=value. The task command
can be specified using command='netstat -an ..'