(TASKS-26) Don't recursively delete
Using `rm -rf` and its powershell equivalent is potentially dangerous if the
temporary directory happens to be invalid. Since we create the temporary
directory and copy the file into it, we can be more selective when deleting. On
Windows we still need to use the `-Force` flag in order to delete read-only or
hidden files.