(TASKS-50) Run the init task when given a module name
If the task name does not contain '::', then it refers to a module, and we
should assume we're running the module's `init` task. The file on disk may have
an extension, e.g. init.sh, ignore the extension if present.

Also fix the exception message so that it includes the original task
name, not the nil value returned from `load_task_file`.