(TASKS-56) Pass task arguments as JSON for powershell
If the task metadata's input_method is 'both' or 'stdin', then send the
task arguments as JSON to each powershell node via stdin.

While the WinRM gem has support for the WriteStdin message[1], the API
does not expose it, and I wasn't able to make it work. Instead wrap
powershell in cmd, and pipe the arguments to powershell.

Note powershell uses backtick instead of backslash as the escape
character. So the task arguments are escaped as prior to piping to
powershell:

    {`"message`":`"hi`"}

[1] https://github.com/WinRb/WinRM/blob/1883937717356756eb10f00d56d8f37e52956cdb/lib/winrm/wsmv/write_stdin.rb#L45