(BOLT-87) Vendor puppet, hiera, facter
Vendor the `feature_puppet-tasks` branch of puppet as a git submodule
and require the puppet functionality only when running a task or plan.

When installing the bolt gem using AIO ruby, the already installed
facter and hiera gems do not satisfy the dependencies that our vendored
puppet requires. It appears the puppet, facter and hiera components of
AIO are not exactly the same as normally installed gems. As a result, gem
install will attempt to install new versions of facter and hiera that
conflict with the already installed versions.

If a user proceeds with installation, the native facter executable will
be overwritten by the ruby facter binstub. To avoid this issue, vendor
facter and hiera in the same way as puppet.

Include the `lib` directory of each vendored submodule when building the
bolt gem.

Include an explicit step for Appveyor to update submodules, this happens
automatically in Travis.

Exclude the `vendored` directory from the rubocop check. However rubocop
has a default set of excludes[1], so we need to exclude the `vendor`
directory in which Travis installs bundles.

[1] https://github.com/bbatsov/rubocop/commit/96a4b4d3e4eb292b323afc62ae3b26cbc89d0cb8#L57-L59