(BOLT-87) Simplify how vendored code is loaded
Previously, we prepended our vendored puppet, facter, hiera lib directories to
the LOAD_PATH each time the `execute`, etc methods were called. In theory, it
could lead to the paths being prepended multiple times.

DRY up the logic by moving the LOAD_PATH manipulation to
`required_vendored.rb`, and have bolt require that relatively. That way
the LOAD_PATH is only modified once. We also only attempt to require the
vendored code in once place.  If we're trying to execute a task or plan,
and the vendored code isn't present, we fail. If it is present, then we
don't need to try to load it again.

Also clarify that `puppet_pal` is intended to be the only entry point into
puppet from bolt. Once BOLT-66 is implemented, we can remove the other require
statements and the call to TaskInformationService.