(BOLT-90) Ensure that plan options includes nodes with non empty strings
Before this commit, the second argument to `run_plan` could be undef,
a hash lacking the 'nodes' entry, or have a 'nodes' entry that contained
empty strings. This is now changed so that a 'nodes' entry is always
present and either an empty array or an array with non-empty strings.