(BOLT-162) Add scaffolding for beaker acceptance
This commit adds basic configuration for running beaker acceptance
tests against `bolt`.

This initial commit only supports testing against an available gem
release and does not include a Rakefile for unifying an interface
for beaker execution.

This setup has been successfully tested on the following platforms:

    * Centos 7 x86_64
    * Ubuntu 16.04 x86_64
    * Windows 10 Enterprise x86_64

Currently, the bolt gem installation passes on OSX 10.12, however due to
[BKR-1215](https://tickets.puppetlabs.com/browse/BKR-1215) the `bolt`
command is not available when the test executes.