Resolve issue with connecting to Vagrant hosts directly
vagrant-ssh always outputs the localhost and port forwarded port, unless
told explicitly to output other information. So we launch with known IPs
so we can set that as ssh.host. We set the port to 22 to avoid the
BOLT-137 issue where bolt doesn't read from the ssh config file for the
port. This requires the BOOT environment variable to work around the
fact vagrant needs to access the machine to configure the network to use
the private IP, but can't use the private network to configure the
private network on first run.