(docs) Update README with submodule command
Previously, the bolt gem would throw an error of `Puppet must be installed to
execute tasks` if Puppet and associated gems were not vendored via git
submodule.  However this error message was misleading as it did not explain
how or where to install Puppet.  This commit updates the README explaining how
to install puppet into bolt.