(BOLT-164) Refer to tasks by name not path
BOLT-164 was caused by bolt preferring to treat the argument to task run
as a path instead of a task name. If you tried to run the 'init' task
from the moduledir bolt would use the relative path to the module as the
task path. This commit removes the ability to run a task by path instead
of by task name to prevent this issue.