(BOLT-152) Add Bolt::ErrorResult
Add a Bolt::Node::BaseError for node related errors. This type of
error is converted to a Bolt::ErrorResult for cases where the node
wants to print an error message, instead of preserving the exception and
stacktrace, like Bolt::ExceptionResult does. The issue_code should be
uniquely identify the exception, suitable for translation, whereas the
kind is the general category of exception.