(BOLT-152) Use secure ssh authentication by default
Previously, bolt would silently accept the remote host's key so long as it
didn't conflict with an existing entry. This is the `lenient` verifier behavior.
While handy for development, it makes bolt susceptible to MITM.

This commit modifies bolt to use the `secure` verifier by default. The
connection will be allowed if the remote host has an entry in our known_hosts
file and the key presented by the host matches our local copy.

Add a command line option (either -k or --insecure) to fallback to the old
lenient behavior. The other net-ssh verifiers `null` and `strict` are not
supported.

Also updates the spec_helper to clear the Bolt.config state before each test so
that we don't have inter-test dependencies.