(BOLT-149) Simplify what files are packaged in the gem
The old approach of `git ls-files` relied on commands which don't
typically exist on Windows, so `bundle install` would fail.

Also we were shipping unnecessary files, like the docs pdf, example
tasks, etc. According to the rubygems specification, `files` should only
contain require-able files[1].

Reduce the set of files that are packaged in the gem to only what is
needed.

[1] http://guides.rubygems.org/specification-reference/#files