Update OSX instructions adding xcode license info
This issue was hard to nail down since I had already run the "xcode-select --install" and was still having the same ruby native extension error that it was supposed to fix. The license issue wasn't very apparent but was the fix for me (seemingly since I otherwise have never opened xcode).