(BOLT-28) Always read password if --password is specified
Previously, it wasn't intuitive that `--password` should be specified
without an argument in order for bolt to prompt for the password. And it
isn't secure to allow users to enter their passwords on the command
line.

This commit changes --password so that if specified and stdin is a tty,
then prompt for the password. If not a tty, then read from stdin, to
enable the password to be piped to bolt.

Also user and password are required for winrm, so print a meaningful error
message if either is missing.