(BOLT-28) Improve help for password prompting
Update help so that it's clear to omit the `--password` value to be
prompted for the password.

Also winrm printed an ugly stack trace if the password is omitted,
because the WinRM connection was created in our initialize method
instead of connect. Move it to the latter so that we gracefully handle
exceptions.