(BOLT-151) Yield node results as they finish
All of the run_* methods now take an optional block. If one is provided, then
the node and its result will be yielded to the block as the operation finishes.
Only one result will be yielded at a time, so callers do not need to worry about
concurrency issues. For example, colorize modifies the global stdout stream
color and doesn't need to worry about concurrency issues.