(BOLT-44) refactor result handling and implement human output
This commit implements the human output format for bolt and adds a
node_started event to outputters to support this.

It also refactors how results are passed simplifying the number of
result classes. Previously there were two sets of result classes one
that was used in the transport for control flow which was translated to
a result of the second type for display. Now there is only a single set
of result classes which are returned by the transport run_* functions
directly. There are no longer separate classes for success and failure
the success? method should be called to differentiate between the two.
Control flow in transport logic has been refactored to use
raise/rescue/ensure instead of result.then/result.ensure.

TODO:
- In order to maintain compatibility with ExecutionResults the to_h
  method of the result class has not been changed. The to_result method
  should be moved to to_h in the future when ExecutionResult can be
  updated
- The transport methods currently catch exceptions and generate "Result"
  instances from them. This is unecessary since Exceptions are caught in
  the executor as well. The transport methods should raise
  Bolt::Node::BaseError subclasses and let the executor handle them