(BOLT-241) Load config from file
This commit adds the ability to load config options from a file. This
required a small refactor to seperate config-like options from
command-options. Config like options are all flags and are passed around
with config but may not be read from the config file. The Command
options are those that are not prefaced with flags and are only exposed
to the execute function of the CLI not passed around with the config.

There is a further distinction between global config options and those
that are transport specific. Transport specific options are specified
for each transport in the config file but only map to a single flag for
all transports as command line flags. Most of these flags only work with
the ssh transport now.