(BOLT-156) Add powershell input method
This commit adds support for the 'powershell' input method to the `Bolt::WinRM`
class.
The implementation injects a base64 encoded JSON blob containing the task
parameter values to the PowerShell session which is then expanded in that
session into powershell's `Hashtable` data structure which is then in turn
used for splatting in the invocation of the task code.