(BOLT-163) Use ruby timeout for WinRM connection timeout
Unfortunately, the combination of read_timeout and operation_timeout
doesn't actually seem to cause any timeout to be applied to the
connection step. This commit changes the WinRM transport to use a Ruby
timeout around the initial connection establishment.

This also increases the default connection timeout to 10 seconds, since 1
seems to be too aggressive for WinRM.