(BOLT-66) Use puppet_pal for task parameters validation
Use the new Puppet PAL (Puppet As a Library) facilities to implement
parameter validation in the `run_task` puppet function provided by the
`boltlib` module. Also update the signatures of the function to
reflect changes in puppet and to cater for the use from the bolt CLI which
is refactored to delegate task execution to this function in a similar
fashion to how it delegates plan execution.