(BOLT-125) Support verified SSL for WinRM connections
We now default to HTTPS connections for WinRM unless --insecure is
specified. This requires a CA certificate to be provided, which can be
done via the winrm ca-cert setting in the config file.

The WinRM connection now uses the appropriate default port depending on
whether it's connecting via HTTP or HTTPS (using ports 5985 and 5986
respectively).