(maint) Avoid spurious test failure due to Ruby bug
There is apparently a bug in older versions of Ruby that causes the
first DNS request made by the tests to take 10 seconds to timeout. Since
the first test we have tries to assert that the connection fails fast
due to DNS failure, and doesn't hit the bolt timeout, this bug causes
the test to spuriously fail. Since the added time seems to only be for
the first test, the simplest way to avoid the problem is to ensure that
the time-sensitive test doesn't run until another test has made a DNS
request and alleviated the issue.