(BOLT-125) Configure SSL for WinRM in vagrant
This adds a cert.pfx file (with both a cert and private key) to the
windows vagrant node and configures it to use that cert for SSL for
WinRM. There's also a corresponding cert.pem (with only the cert itself)
for bolt to use to verify connections to the node.

These files are stored in the repo rather than being generated during
vagrant provisioning because the windows host is Windows Nano, which
doesn't have the necessary New-SelfSignedCertificate command to generate
the cert.