(BOLT-266) Refactor logging
This centralizers the logger config state with the @@config class
variable on the Logger class. Classes that need a logger should get one
with Logger.get_logger but be careful not to persist that logger unless
configuration is finished.

This also adds more executor level logging and varies the verbosity of
that logging depending on whether the executor is running inside a plan
or not. This allows more verbose logging during plan execution without
duplicating log messages and output during other commands. This is a
temporary solution while we experiment with plan output