(BOLT-81) implement disable_ssh_agent option
Prior to this commit, bolt reported an error
when Net-SSH attempted to use an unconfigured ssh-agent.

Net-SSH offers a use_agent option which can be used to disable ssh-agent.

With this commit, bolt offers a disable-ssh-agent option
which can be used to used to specify the Net-SSH use_agent option.