(BOLT-81) check for ssh_agent
Prior to this commit, bolt reported an error when Net-SSH
attempted to use an unconfigured ssh-agent or peagent.

Net-SSH offers a use_agent option which, when false, will not use an agent.
Bolt sets use_agent to true by default.

With this commit, bolt checks for ssh-agent or peagent to set use_agent.