(BOLT-268) Switch from built-in Logger to Logging gem
This commit swaps out the built-in Ruby Logger class for the Logging
gem, modeled after log4j.

This allows us much better control over which components log at which
levels, as well as allowing multiple log destinations (ie. both STDERR
and a file). It also includes built-in support for problems like
colorization that we would otherwise need to roll on our own.