(BOLT-286) Convert ExecutionResult to ResultSet
This standardizes how results are handled in bolt and puppet. Previously
a hash Bolt::Result objects were generated by they executor when a
run_* action was taken from puppet. The puppet run_* functions then
convverted those into and ExecutionResult. Now Results are built into a
Bolt::ResultSet object in the executor. Both this and the bolt result
are now pcore datatypes and available directly in the plan.

This adds the Bolt::PAL class which is only used for testing so far. In
the near future all interaction with Puppet should go through this
class.