(BOLT-209) Return exitcode 2 when some nodes fail an operation
Previously, as long as an operation was started successfully, bolt would
always exit 0, regardless of whether it actually completed successfully.
This meant that it was impossible to tell whether a task or command had
actually run on the nodes in question, only whether the task exists
or whether the command was properly interpreted by bolt.

For tasks, scripts, comamnds and file uploads, we now exit 2 if any node
fails. The behavior for plans is unchanged, as we still don't have a
solid definition of a plan "result". It can now be assumed that if bolt
exits 0, the operation completed successfully on every node.