(BOLT-277) Provide pointers for common WinRM connection error cases
Bolt defaults to using SSL for WinRM unless explicitly instructed
otherwise, but the default configuration for Windows hosts is to not
accept SSL connections for WinRM. Because of that, a common failure
scenario is to try to run bolt against a node and have it fail to
connect because the SSL port isn't open. Since we want to maintain a
secure default, this commit adds logic to detect that case and mention
what the problem might be, pointing users to try --insecure.

Another potential error case is trying to use SSL to connect to the
non-SSL port, which result in an extremely arcane OpenSSL error:

    SSL_connect returned=1 errno=0 state=SSLv2/v3 read server hello A: unknown protocol

We now also detect this scenario and mention the likely cause.