(BOLT-304) Add support for displaying plan parameters
This commit adds support for displaying plan parameters info using:
  bolt plan show <plan name>
A new function for printing the info - `print_plan_info` - is added to
the outputter classes. The existing `print_plan` function is renamed to
`print_plan_result` so as to distinguish it better from the newly added
`print_plan_info`.