(maint) Fix Vagrant setup for winrm over ssl tests
Adds a command to the Vagrant initialization script of the Windows VM used
for the winrm tests to open the winrm over ssl port in the Windows firewall.
Additionaly replaces the self signed certificate used for winrm over ssl
tests with a CA & client certificate pair as some ssl implementations refuse
to verify self signed certificates. The `resources/cert.pem` file is renamed
to `resources/ca.pem` to make it clearer what the file is.
Since these certificates are also used in AppVeyor, the `appveyor.yml` is
updated accordingly too.