(maint) Update submodules to pull in fixes for listing tasks/plans
Previously, listing tasks or plans with `bolt task show` or `bolt plan
show` would fail if any task or plan was invalid. This issue was fixed
in Puppet, and this commit updates the submodule to a version that
includes the fix. This required a couple of test updates to account for
the new and improved behavior.