(BOLT-218) Infer SSH user from SSH config or current login
Previously, if a username wasn't specified either as part of the target
or from the CLI/config, Bolt didn't know what user the node was going
to eventually connect as. We now replicate the logic that net-ssh uses
to determine the user, by checking the SSH config and falling back to
the currently logged in user. This will allow us to reliably make
decisions (for instance, whether we need to sudo) based on who we're
connecting as.