(BOLT-218) Allow setting run_as for individual execute invocations
Previously, the `execute` function would *only* honor the @run_as
variable, which makes complex behaviors involving multiple users
difficult. It now accepts a `run_as` argument to allow the user to be
explicitly specified even if we're in the middle of a high-level
operation that is meant to be done as a different user. In particular,
this lets us explicitly run commands as root when needed, even if
neither the user we're connecting as or "running as" is root.