(BOLT-218) Honor "run_as" for file uploads
Previously, `bolt file upload` would leave the owner of uploaded files
as whichever user it connected as, rather than the user set by "run_as".
That led to situations where a file might be uploaded and unreadable by
the user it's intended to be used by.

We now upload the file to a secure tempdir, change the owner to the
correct user, and then move it into its correct location. This avoids
the user we're uploading the file as from ever being able to read the
file.

Because file ownership can only be transferred by root, we now
explicitly require root access in order to upload as one non-root user
and run as another.