(BOLT-218) Upload tasks/scripts as the run-as user
Previously, uploaded tasks and scripts would be owned by the user we
connected as rather than the user we're running as. That meant that
whether or not the task could actually be run depended on what its
original permissions were on the local system.

We now handle this properly by uploading the file to a tempdir and then
chown-ing the entire tempdir to the run-as user. This ensures that the
file will be correctly executable.