(BOLT-255) Create Transport classes that accept Targets
This commit removes the previous Node classes which represented both a
host and its connection. We now use stateless Transport classes which
take Targets are arguments.

In this commit, all of the actual work is done by a connection object
specific to each transport. The purpose of the transport is to handle
making a connection out of the target and then delegating the work.

Transports are now responsible for managing the lifecycle of a
connection, rather than the executor. The actual lifecycle remains
unchanged; a connection is only open for the lifetime of a single
task/command/etc.