(BOLT-316) Reuse threadpool between executions
This commit changes from a FixedThreadPool to a CachedThreadPool and
allows it to be reused between executions. Whereas previously we
submitted all the work to the pool and then shut it down to determine
when the work was done, we now use futures to perform the work, so that
we can simply block awaiting the values of the futures.

This allows us to avoid having to shutdown the threadpool, so we can
keep it around for future invocations.