(BOLT-255) Don't modify target object for run_as
This introduces a clear separation between the use of run_as on the
target and run_as on the connection. The target's run_as value
represents the *default* user to perform operations as, whereas the
connection can be overridden for a specific operation. This cleans up
the lifecycle of that value and avoids mutating the target.