(BOLT-318) Move nodes list splitting to `get_targets`
Moves splitting a comma-separated node list to `Inventory#get_targets`.
One implication is that we will now split on commas for node names
passed in a plan as well as on the CLI. That simplifies passing lists of
names to a plan, but also prevents passing things with commas (such as
passwords) on target URIs. The config section of the inventory file can
be used instead.

For example, the following should now work:
```
plan example(String $targets) {
  run_task('example::task', $targets)
}
```
then run `bolt plan run example targets=a,b,c`.

Loading the inventory and splitting `--nodes` from the CLI is moved to
`CLI#parse` to simplify testing and move error-checking earlier.