(BOLT-316) Batch task invocations
This commit implements a batch_task() method on transports, which is
used to indicate to a transport that it should run multiple tasks as
part of a single batch. This is specifically useful for the orch
transport, which can run many nodes as part of a single orchestrator
job.

Transports now have the option of either inheriting from
Transport::Base, which will provide a default batch implementation that
just calls run_task() for every node, or implementing batch_task()
directly and controlling their own concurrency.