(BOLT-342) Add TargetSpec type alias to replace TargetOrTargets
This moves the TargetOrTargets type alias out of the boltlib functions
and directly into the types/ directory of boltlib, as the new
Boltlib::TargetSpec alias. We also create a top-level TargetSpec alias
that refers to Boltlib::TargetSpec, so that users who wish to specify
that type in their plans can do it without having to namespace it under
Boltlib, which is purely an implementation detail.

The TargetSpec type accepts a String, a Target, or any set of Arrays of
those two types.

Specifically, TargetSpec represents the set of types that can be passed
to get_targets() and result in an Array[Target]. For the most part,
this type should be transparent to users, since it can be passed through
to other plans and to run_task, etc. In the case where users need to do
some sort of logic based on the *actual* Target objects, get_targets()
can be used to "resolve" the TargetSpec to a list.