(BOLT-331) Add a command to generate inventory from PuppetDB
This adds the bolt-inventory-pdb command which can be used to populate a
"template" inventory file with nodes based on PuppetDB queries.

The input to the command is basically an inventory file, except groups
should have a "query" field instead of a "nodes" field. The command will
read the input, execute each of the queries against PuppetDB, and print
a complete inventory with the "nodes" field populated for each group.

This command supports both token and cert auth to PuppetDB, and will
read the puppetdb.conf from the `puppet query` command if it's present.