(BOLT-263) Add support for logging into file(s)
Adds an option to configure logging into file(s). The log files are defined
in the bolt's configuration file 'log' entry which is a mapping (hash) of
file paths of the log files (optionally prefixed with the 'file:' schema)
to hashes of parameters of the respective logs. The parameter hashes may
include:
 - level - the log level of the log file
 - append - a flag specifying whether the log should be appended to or rather
            overwritten
A special log file named 'console' may be used to override the log level of
the default console logger.

Example:
...
log:
  console:
    level: warn
  file:~/.bolt/bolt.log:
  file:~/.bolt/debug.log:
    level: debug
    append: false
...