(maint) Use bolt user for tests
Previously we used the vagrant user when running tests against Vagrant managed
VMs.
This patch changes that to use the bolt user instead. The Vagrantfile is
updated to ensure the appropriate user account is created on the VMs.
Since some of the config is shared with AppVeyor, the appveyor.yml is updated
accordingly too.