(maint) Properly construct error result of last resort
This code tries to ensure that every Promise we construct will get a
value, no matter what happens during execution, in order to avoid a
deadlock. However, the call to create the error result itself had a bug,
causing... a deadlock. We now properly pass the target when creating the
error result, avoiding aforementioned deadlock.