(BOLT-336) Treat --nodes as the nodes parameter for run plan
This change makes it possible to specify the nodes parameter of a plan by
passing the --nodes option(s). (If more than one is passed then the values
they specify are concatenated comma delimited.)
It is an error to specify both the --nodes options and the nodes parameter.
It is also an error to specify the --nodes option when running a plan which
doesn't take the nodes parameter. The specified value *is* however passed
to the plan as the nodes parameter, but that fails as that parameter is not
expected by the plan.
While the use of the --nodes option in the way enabled by this commit is
allowed it is not encouraged and so the option is not printed in the help
messages related to 'bolt plan' commands produced by 'bolt help'. This
necessitated creation of a subclass of the OptionParser used for producing
these messages with support for excluding selected options from those
messages on demand.