(perf) Improve performance when searching a large inventory.
This converts @nodes in an inventory group to a Hash so it can be more
quickly searched. In practice this reduces the time of a
get_targets call for 50k targets and a small number of groups from minutes to <10s.

Depending on the real world structure of large inventory files there may
be more improvements to make here. The biggest problem right now is
repeatedly updating targets with every get_targets call though.