(BOLT-352) Fix inventory file config to match config file format
Fixes the inventory file config section so it uses structure and keys
that are consistent with the config file format. Previously it required
them to match the internal structure of the Bolt::Config object, which
led to a lot of inconsistency between inventory and config file formats.

So
```
config:
  transports:
    ssh:
      host_key_check: false
```
is now
```
config:
  ssh:
    host-key-check: false
```