(BOLT-167) Query PuppetDB for targets with --query flag
This adds a --query flag to specify a PQL query to use to determine
targets. The query will be executed against the PuppetDB instance
configured in ~/.puppetlabs/client-tools/puppetdb.conf (thereby working
automatically for users who have PE client tools configured). In order
to serve as targets, the query results must include a 'certname' field,
which will be used as the target name. Targets are then looked up in the
inventory to find connection information.

This flag is mutually exclusive with the --nodes flag. Unlike --nodes,
it can only be specified once.