(BOLT-356) Add pcp/local-validation setting
Adds support for the pcp/local-validation configuration setting.
When set to `false` then, under the condition that a task is run on targets
exclusively using the `pcp` transport, no validation of the specified task
parameters against the local task definition is performed (the definition
is not even loaded so the task doesn't even need to be installed locally).
The validation still takes place if the task is run on targets which use
different transports (which may or may not include the `pcp` transport).