(PUP-8441) Serialize pcore values in plans
Currently plans are allowed to return arbitrary objects which may not
support to_json and therefore will error when we try to serialize them.
Worse, pcore objects tend to blow up when to_s is called on them outside
a compiler.

We should choose between supporting arbitrary pcore objects as the
return value for plans and handling serialization of them or limiting
the return to datatypes we expect. Currentlty I think the valid types
are Data, Error, Result, ResultSet, Target, and combinations thereof.

If we do decide on serializing pcore we need to choose between
displaying it as raw json like we do with the expected types of using
the rich data format.

I personally lean towards restricting the return types of plans. I think
it's adaquate for now and it's easier to expand the spec to support more
return types than restrict it in the future.

Notes about this implementation:

I'm recreating the ruby objects we care about so that we can handle them
more cleanly outside of pal. An alternate implementation would be to
keep the compiler arround for duration of a bolt execution and serialize
pcore more later as needed.

f