(BOLT-135) Select task implementation based on available features
This commit adds logic to the existing transports to dynamically select
the appropriate task implementation based on the available features of a
target. Target now has a select_impl() method which, given a task and an
optional list of "additional" features, will return the path to the
first suitable executable. The "additional" features are passed by the
transport and account for any features which are directly *implied* to
be present based on the transport (ie. 'shell' for ssh and 'powershell'
for winrm).

This allows a single task to have multiple implementations for different
platforms and for Bolt to choose the appropriate one for each target at
runtime.