(maint) Use bolt inventory during acceptance tests
This replaces the boilerplate setup in many tests with a single
inventory file that has SSH and WinRM groups defined with the
appropriate config included. This lets the tests just refer to the
"ssh_nodes" or "winrm_nodes" groups, without having to worry about host
key checking, SSL, or username/password.