(maint) Avoid environment-specific test failure
This test was trying to check that 'localhost' is properly overridden by
an entry in the inventory file, by checking that connecting over SSH
causes a connection error. However, if the default config on the node
running the test *allows* connection to localhost, this test will fail.

We now ensure that this test tries to connect to a closed port, to avoid
any environment-specific issues.