(BOLT-439) Support inventory groups defined within other groups
Previously, the code needed to support nested inventory groups was
present, but with an explicit check to disable it. This commit removes
that check and fixes a bug where variables were merged in reverse order
compared to other values.

Groups can now be defined within other groups. Variables/config prefer the
first "branch" (top-level group) that the node matches, and use the most
deeply nested (most specific) group within that branch. In other words,
we use depth-first lookup, and merge across branches.

A group is not allowed to be defined as a child of multiple parent
groups. This triggers the same error as redefining a group did
previously.