(BOLT-372) Send information to orchestrator when running a plan
When a plan is run the orchestrator transport will send the plan_start
command the first time a task runs on a node through that orchestrator.
Every task will be run with the plan_task command instead of the task
command and before exiting bolt will send a plan_finish command with the
result to each orchestrator that was used during the plan.