(BOLT-455) Use bolt logger from puppet
This adds a stub 'logging' logdestination type to Puppet that passes log
messages through to the Logging logger used by bolt. This causes
Puppet's logs to be logged alongside bolt logs, with the same format.
This also means that Puppet logs properly get logged to files in
addition to the console.

This also fixes a bug where we weren't actually logging in Puppet,
because `with_puppet_settings` was resetting the log destinations to
empty. We now reconfigure logging every time we re-initialize settings,
so that we always log.

We now use the 'debug' log level for Puppet, causing it to "log"
*everything* (which really means pass it through to the Logging logger),
allowing the Logging logger to decide which messages to actually log.