(BOLT-453) Print only log messages on the console
Previously, we used a fairly overbearing log format for the console,
including the complete ISO-8601 date time with microsecond-precision,
the name of the class doing the logging, and the log level. We now print
only the message itself, excluding the timestamp, logger name and the
log level.

We now use colors to indicate the log level on the console, with red
representing errors, yellow for warnings, green for notices and the rest
uncolored.

Because the context can sometimes be useful, we continue to include it
in the log format we use for log files. We also exclude colors for log
files.