(BOLT-453) Include duration after running an action
Previously, we included timestamps with every log message, which was
useful for understanding how long a given step of a plan took. Since we
no longer show timestamps, we now just include the duration in seconds
directly in the "finished" message.

We also now only list the nodes targeted by an action if there are 5 or
fewer; otherwise we just show a count. Because start and finish messages
are directly correlated, we also no longer repeat the node count in the
finished message.